A-starter-python/requirements.txt
flask~=3.0.0
SQLAlchemy
grpcio-tools==1.64.1
grpcio
protobuf==3.20.*
py_eureka_client
amqpstorm
argparse
pymongo
pyyaml
grpc-interceptor
asyncio
fastapi[uvicorn]
ariadne
uvicorn